designspace_extractor/requirements.txt
# Core dependencies
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Code parsing
ast-comments>=1.1.0

# MATLAB/Scientific computing
scipy>=1.10.0
h5py>=3.8.0

# PDF extraction
grobid-client-python>=0.0.5
pypdf>=3.0.0

# Google Sheets integration
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# LLM integration
anthropic>=0.18.0
openai>=1.0.0

# JavaScript parsing (via Node.js subprocess)
# Requires: npm install -g @babel/parser @babel/traverse

# HED validation
hedtools>=0.4.0

# CLI & utilities
click>=8.1.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0

# Testing & validation
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
